pandas2_examples/pyproject.toml
[tool.poetry]
name = "pandas2-examples"
version = "0.1.0"
description = ""
authors = ["Axel Furlan <axel.furlan@toptal.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "3.8.6"
pandas = "v2.0.0"
google-cloud-bigquery = "^3.11.4"
pyarrow = "^13.0.0"
sqlalchemy = "1.4.49"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
db-dtypes = "^1.1.1"
sqlalchemy-bigquery = "^1.8.0"
pandas-gbq = "^0.19.2"
gcsfs = "^2023.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
